pyproject.toml
[tool.poetry]
name = "classroom"
version = "0.1.0"
description = "Classroom Rest API"
authors = ["naresh-khatri <naresh.khatri2345@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.2"
djangorestframework = "^3.12.4"
django-allauth = "^0.44.0"
dj-rest-auth = "^2.1.4"
django-cors-headers = "^3.7.0"
django-filter = "^2.4.0"
djangorestframework-simplejwt = "^4.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
